requirements.txt
sqlmodel
fastapi[standard]
asyncpg
alembic
pydantic-settings
python-multipart
pytest
httpx
ruff
email-validator
bcrypt==4.0.1
python-jose[cryptography]
google-auth-oauthlib 
google-api-python-client